+++ R/transformed.R
@@ -106,6 +106,7 @@ density.dist_transformed <- function(x, at, deriv_method = "symbolic",
 cdf.dist_transformed <- function(x, q, ...){
   inv <- function(v) suppressWarnings(x[["inverse"]](v))
   p <- cdf(x[["dist"]], inv(q), ...)
+  if(!monotonic_increasing(x[["transform"]], support(x[["dist"]]))) p <- 1 - p
   limits <- field(support(x), "lim")[[1]]
   if (!any(is.na(limits))) {
     p[q <= limits[1]] <- 0
@@ -116,6 +117,7 @@ cdf.dist_transformed <- function(x, q, ...){
 
 #' @export
 quantile.dist_transformed <- function(x, p, ...){
+  if(!monotonic_increasing(x[["transform"]], support(x[["dist"]]))) p <- 1 - p
   x[["transform"]](quantile(x[["dist"]], p, ...))
 }
 
@@ -147,3 +149,8 @@ covariance.dist_transformed <- function(x, ...){
   )
 }
 
+monotonic_increasing <- function(f, support) {
+  # Currently assumes (without checking, #9) monotonicity of f over the domain
+  x <- f(field(support, "lim")[[1]])
+  x[[2L]] > x[[1L]]
+}